+++ src/main/java/cart/application/CartItemService.java
@@ -2,15 +2,10 @@ package cart.application;
 
 import cart.application.dto.request.CartItemQuantityUpdateRequest;
 import cart.application.dto.request.CartItemRequest;
-import cart.application.dto.request.PaymentRequest;
 import cart.application.dto.response.CartItemResponse;
 import cart.domain.cart.CartItem;
 import cart.domain.member.Member;
-import cart.domain.order.Order;
-import cart.domain.order.OrderProducts;
 import cart.domain.product.Product;
-import cart.persistence.dao.OrderHistoryDao;
-import cart.persistence.dao.OrderProductDao;
 import cart.persistence.repository.CartItemRepository;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
@@ -22,13 +17,9 @@ import java.util.stream.Collectors;
 public class CartItemService {
 
     private final CartItemRepository cartItemRepository;
-    private final OrderHistoryDao orderHistoryDao;
-    private final OrderProductDao orderProductDao;
 
-    public CartItemService(final CartItemRepository cartItemRepository, final OrderHistoryDao orderHistoryDao, final OrderProductDao orderProductDao) {
+    public CartItemService(final CartItemRepository cartItemRepository) {
         this.cartItemRepository = cartItemRepository;
-        this.orderHistoryDao = orderHistoryDao;
-        this.orderProductDao = orderProductDao;
     }
 
     @Transactional(readOnly = true)
@@ -65,31 +56,5 @@ public class CartItemService {
         cartItem.checkOwner(member);
 
         cartItemRepository.deleteCartItemById(id);
-    }
-
-    @Transactional
-    public Long payment(final Member member, final PaymentRequest request) {
-        final List<Long> cartIds = extractCartIds(request);
-        final List<CartItem> cartItems = getCartItems(member, cartIds);
-        member.usePoint(request.getPoint());
-        final OrderProducts orderProducts = new OrderProducts(cartItems);
-        final Order order = new Order(null, member, request.getPoint(), orderProducts);
-        final Long orderId = orderHistoryDao.createOrder(order);
-        member.savePoint(order.getSavedPoint());
-        orderProductDao.createProducts(orderId, order.getOrderProducts());
-        cartItemRepository.updatePoint(member);
-        return orderId;
-    }
-
-    private List<Long> extractCartIds(final PaymentRequest request) {
-        return request.getCartItemRequests().stream()
-                .map(CartItemRequest::getProductId)
-                .collect(Collectors.toList());
-    }
-
-    private List<CartItem> getCartItems(final Member member, final List<Long> cartIds) {
-        final List<CartItem> cartItems = cartItemRepository.getCartItemsByIds(cartIds);
-        cartItems.forEach(cartItem -> cartItem.checkOwner(member));
-        return cartItems;
     }
 }
